+++ client/src/components/account/options/Security.vue
@@ -19,17 +19,15 @@ export default {
 <style scoped>
 
 button {
-  width: auto;
-  height: auto;
-  padding: 0.5rem 1.5rem;
-  margin: 1rem;
-  text-wrap: nowrap;
-  font-size: 1.2rem;
+  width: 12%;
+  height: 7%;
+  margin: 1.5rem;
+  font-size: 1.4rem;
   font-family: 'Poppins', sans-serif;
-  background-color: #FFD586;
+  background-color: rgba(240, 128, 128, 0.55);
   border-radius: 0.6rem;
-  box-shadow: -0.05rem 0.05rem 0.8rem 0 rgba(0, 0, 0, 0.25);
-  border: 0.1rem solid black;
+  box-shadow: 0.01rem 0.01rem 0.5rem 0.1rem rgba(0, 0, 0, 0.2);
+  border: none;
   cursor: pointer;
   transition: background-color 0.3s ease,
   border-radius 0.3s ease;
@@ -39,4 +37,5 @@ button:hover {
   background-color: #FFC94D;
   border-radius: 1.5rem;
 }
+
 </style>