+++ client/src/components/home/HomeHeader.vue
@@ -6,7 +6,7 @@
         <li @click = "selectTicketClass(ticketClass)" v-for = "(ticketClass, index) in ticketClasses" :key = "index" :class = "{selected : selectedTicketClass === ticketClass}" class = "no-select">{{ ticketClass }}</li>
       </ul>
     </div>
-    <button @click = "account" class = "no-select">My Account</button>
+    <img id = "account" src = "../../assets/account.png" @click = "account" class = "no-select">
   </header>
 </template>
 
@@ -15,7 +15,7 @@ export default{
   name: "HomeHeader",
   data() {
     return {
-      ticketClasses: ["Movies", "Shows", "Sports", "Travels", "Virtual"],
+      ticketClasses: ["Movies", "Shows", "Sports", "Travels"],
       selectedTicketClass: "Movies"
     }
   },
@@ -36,34 +36,40 @@ export default{
 
 img {
   padding-top: 0.5rem;
-  padding-left: 1rem;
-  width: min(13.25%, 21vh);
+  padding-left: 2rem;
+  width: min(16%, 14vw);
   height: auto;
   margin-right: 1rem;
   cursor: pointer;
 }
 
+#account {
+  width: 2.5%;
+  margin: 0 1.5rem 1rem 0
+}
+
 li {
   display: inline-block;
   list-style-type: none;
-  font-size: 1.3rem;
+  font-size: 1.1rem;
   margin-right: 1.2rem;
   padding: 0.4rem 1.2rem;
   color: #333;
   cursor: pointer;
-  border-radius: 0.8rem;
   transition: background-color 0.5s ease,
-              box-shadow 0.25s ease;
+              border-radius 0.5s ease,
+              box-shadow 0.5s ease;
 }
 
 li:hover {
-  background-color: #F6ACAC;
+  background-color: rgba(255, 201, 77, 0.7);
+  border-radius: 0.6rem;
 }
 
 .selected, .selected:hover {
-  background-color: #E26868;
-  border: 0.1rem solid #000;
-  box-shadow: -0.05rem 0.05rem 0.8rem 0 rgba(0, 0, 0, 0.25);
+  background-color: #FFC94D;
+  border-radius: 0.6rem;
+  box-shadow: -0.01rem 0.01rem 0.8rem 0 rgba(0, 0, 0, 0.1);
 }
 
 </style>