+++ client/src/components/home/HomePage.vue
@@ -47,7 +47,7 @@ export default{
   display: flex;
   justify-content: center;
   width: 100%;
-  background: linear-gradient(to right, #DFDFDE, #F5EDDC);
+  background: #FFF;
 }
 
 </style>