+++ client/src/routes/index.js
@@ -3,6 +3,7 @@ import MyAccount from "@/components/account/MyAccount.vue";
 import ResetPassword from "../components/authentication/forms/ResetPassword.vue";
 import HomePage from "../components/home/HomePage.vue";
 import VerificationError from "../components/authentication/VerificationError.vue";
+import Booking from "@/components/events/Booking.vue";
 
 export default [
     {
@@ -18,8 +19,12 @@ export default [
         component: ResetPassword,
     },
     {
-        path: '/home',
+        path: '/events',
         component: HomePage,
+    },
+    {
+        path: '/booking',
+        component: Booking,
     },
     {
         path: '/verification-error',