+++ server/routes/eventRoutes.js
@@ -6,4 +6,10 @@ const router = express.Router();
 // Get Movies
 router.get('/api/events/movies', controller.getMovies);
 
+// Add to Favorites (Movies)
+router.post('/api/events/movies/favorites', controller.addFavorites)
+
+// Get Favorites
+router.post('/api/account/favorites', controller.getFavorites)
+
 module.exports = router;